lib: Move tests to appropriate files.
